Desarrollo/SVO/Fuente/SVO_CF/accounts/views.py
from rest_framework import status
from rest_framework.generics import CreateAPIView
from rest_framework.permissions import AllowAny
from rest_framework.response import Response
from accounts.serializers import CreateUserSerializer


class CreateUserAPI(CreateAPIView):
    """ Registro de usuario
    """
    serializer_class = CreateUserSerializer
    permission_classes = (AllowAny,)
    authentication_classes = ()

    def create(self, request, *args, **kwargs):
        serializer = self.get_serializer(data=request.data)
        serializer.is_valid(raise_exception=True)
        self.perform_create(serializer)
        headers = self.get_success_headers(serializer.data)
        return Response(status=status.HTTP_201_CREATED, headers=headers)
